Split some of the option functions from the main app file.
