+++ src/js/app.ts
@@ -1,15 +1,7 @@
-import 'regenerator-runtime/runtime'
-import { testPuzzle } from "./test-puzzle";
-import { pocketPuzzle } from "./pocket-puzzle";
-import { greatPuzzle } from "./great-puzzle";
-import { PuzzleDefinition } from "./common-data-schema";
+import { getSelector, getPuzzleType, getTileSelection, placeTile } from "./app-options";
 import { DisplayManager } from "./puzzle-display";
 import { getTetrahedron, getTilePool } from "./puzzle-loader";
 import { Tetrahedron } from "./tetrahedron";
-import { Tile } from "./tile";
-import { TilePool } from "./tile-pool";
-import { TilePosition } from "./tile-position";
-import { getRandomInt } from "./utils";
 
 
 // Track tile placing event timer.
@@ -32,78 +24,6 @@ function attachRotateEvents(puzzleDisplay: HTMLElement, tetrahedron: Tetrahedron
                 });
             }
         });
-}
-
-function getSelector(name: string): string {
-    const selection  = <NodeListOf<HTMLInputElement>>document.querySelectorAll(`input[name = "${name}"]`)!;
-    for (const rb of selection) {
-        if (rb.checked) {
-            return rb.value;
-        }
-    }
-    throw new Error("Expected radio option to be selected!");
-}
-
-function getPuzzleType(): PuzzleDefinition {
-    const puzzleType = getSelector("puzzle-type");
-    switch (puzzleType) {
-        case "Test":
-            return testPuzzle;
-        case "Pocket":
-            return pocketPuzzle;
-        case "Great":
-            return greatPuzzle;
-        default:
-            throw new Error("Invalid puzzle type option!");
-    }
-}
-
-function getTileSelection(tilePool: TilePool): Tile | null {
-    const selection = getSelector("tile-selection");
-    switch (selection) {
-        case "Random":
-            return tilePool.randomTile;
-        case "Sequential":
-            return tilePool.nextTile;
-        case "Test":
-            return tilePool.testTile;
-        default:
-            throw new Error("Invalid tile selection option!");
-    }
-}
-
-function rotateTile(tilePosition: TilePosition): TilePosition {
-    const rotation = getSelector("tile-rotation");
-    switch (rotation) {
-        case "None":
-            return tilePosition;
-        case "Random":
-            for (let i = getRandomInt(3); i > 0; --i) {
-                tilePosition.rotateTile();
-            }
-            return tilePosition;
-        default:
-            throw new Error("Invalid tile rotation option!");
-    }
-}
-
-function placeTile(tile: Tile, tetrahedron: Tetrahedron): TilePosition  {
-    const placement = getSelector("tile-placement");
-    let tilePlacedPosition;
-    switch (placement) {
-        case "Random":
-            tilePlacedPosition = tetrahedron.placeTileRandomly(tile);
-            break;
-        case "Sequential":
-            tilePlacedPosition = tetrahedron.placeTileSequentially(tile);
-            break
-        default:
-            throw new Error("Invalid tile placement option!");
-    }
-    if (!tilePlacedPosition) {
-        throw new Error("Failed to place tile on puzzle!");
-    }
-    return rotateTile(tilePlacedPosition);
 }
 
 function testDisplay(): void {
